% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysqrt.R
\name{mysqrt}
\alias{mysqrt}
\title{Home-made square-root function}
\usage{
mysqrt(x)
}
\arguments{
\item{x}{A vector of numerical values.}
}
\value{
A vector containing the square root of the input.
}
\description{
A light wrapper that does nothing special except calling \code{sqrt()} itself.
}
\examples{
mysqrt(4)
mysqrt(3)
mysqrt(1:10)
mysqrt(-1)

}
